ai-security-assistant-backend/requirements.txt
fastapi
uvicorn
pydantic>=2
httpx
openai>=1.52.0
dnspython
python-dotenv
